backend/requirements.txt
fastapi[standard]>=0.110.0
uvicorn>=0.27.1
sqlalchemy>=2.0.27
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
alembic>=1.13.1
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
boto3>=1.34.50
google-generativeai>=0.4.0
pydantic>=2.6.2
pydantic-settings>=2.2.1
PyPDF2>=3.0.1
pdfplumber>=0.10.4
python-dotenv>=1.0.1
pytest>=8.0.2
httpx>=0.27.0
